requirement.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
mccabe==0.6.1
Pillow==8.4.0
psycopg-binary==3.0.4
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.7
zipp==3.6.0
